Make list group generic
