--- Buform/Items/List/IListFormItem.cs
@@ -1,6 +1,0 @@
-namespace Buform;
-
-public interface IListFormItem : IFormItem
-{
-    string? FormattedValue { get; }
-}
