+++ Buform/Items/Text/IMultilineTextInputFormItem.cs
@@ -1,0 +1,12 @@
+namespace Buform;
+
+public interface IMultilineTextInputFormItem : IValidatableFormItem
+{
+    string? Placeholder { get; }
+
+    TextInputType InputType { get; }
+
+    string? FormattedValue { get; }
+
+    void SetValue(string? value);
+}
