--- Buform/Platforms/Ios/Items/List/ListFormItemComponent.cs
@@ -1,12 +1,0 @@
-namespace Buform;
-
-[Preserve(AllMembers = true)]
-[FormComponent]
-// ReSharper disable once UnusedType.Global
-public sealed class ListFormItemComponent : IFormComponent
-{
-    public void Register()
-    {
-        FormPlatform.RegisterItemClass<IListFormItem, ListFormCell>();
-    }
-}
