+++ Buform/Platforms/Ios/Items/Text/MultilineTextInputFormCell.cs
@@ -1,0 +1,20 @@
+using ObjCRuntime;
+
+namespace Buform;
+
+[Preserve(AllMembers = true)]
+[Register(nameof(MultilineTextInputFormCell))]
+public sealed class MultilineTextInputFormCell
+    : MultilineTextInputFormCellBase<IMultilineTextInputFormItem>
+{
+    public MultilineTextInputFormCell()
+    {
+        /* Required constructor */
+    }
+
+    public MultilineTextInputFormCell(NativeHandle handle)
+        : base(handle)
+    {
+        /* Required constructor */
+    }
+}
