+++ Buform/Platforms/Ios/Items/Text/TextFormItemComponent.cs
@@ -7,7 +7,8 @@ public sealed class TextFormItemComponent : IFormComponent
 {
     public void Register()
     {
-        FormPlatform.RegisterItemClass<IMultilineTextFormItem, MultilineTextFormCell>();
+        FormPlatform.RegisterItemClass<IMultilineTextInputFormItem, MultilineTextInputFormCell>();
+        FormPlatform.RegisterItemClass<ITextInputFormItem, TextInputFormCell>();
         FormPlatform.RegisterItemClass<ITextFormItem, TextFormCell>();
     }
 }
